+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1. Adapt block related methods to include all tx from one Epoch
 2. Support EIP-1898
 3. Pass through `eth_sendRawTransaction` params
+4. Fix `eth_getTransactionCount`
 
 ## v0.5.0